requirements.txt
gradio
opencv-python
mediapipe
numpy
Pillow
fastapi
uvicorn
